chore: move _backend to function sub-module
